target encodingのbugfix

+++ analysis_for_kaggle/modules/TableDataTranslater.py
@@ -83,7 +83,7 @@ class TableDataTranslater(CommonDataTranslater):
                 # add Laplace Noize for not data leak
                 # np.random.seed(seed=42)
                 # transed += [np.random.laplace() for _ in range(len(transed))]
-                transed = transed.reshape(-1, 1)
+                transed = transed.reshape(-1, 1).astype(float)
             else:
                 logger.error('NOT IMPLEMENTED CATEGORIZE: %s' % model)
                 raise Exception('NOT IMPLEMENTED')
